requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Explainable AI
captum>=0.6.0
shap>=0.42.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
opencv-python-headless>=4.8.0
albumentations>=1.3.0
scikit-learn>=1.3.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Interface
streamlit>=1.25.0

# Utilities
tqdm>=4.65.0
PyYAML>=6.0